fix : 입력 기능 분리

+++ src/main/java/com/prgrms/vouchermanager/exception/NotCorrectScopeException.java
@@ -4,7 +4,7 @@ import com.prgrms.vouchermanager.message.ConsoleMessage;
 import com.prgrms.vouchermanager.message.LogMessage;
 
 public class NotCorrectScopeException extends MyException {
-    public NotCorrectScopeException(Long discount) {
+    public NotCorrectScopeException(int discount) {
         super(String.format(LogMessage.NOT_CORRECT_SCOPE.getMessage(), discount));
     }
